+++ lib/classes/class-bootstrap.php
@@ -48,8 +48,6 @@ namespace UsabilityDynamics\WPP {
         include_once $this->path( 'lib/class_functions.php', 'dir' );
         /** Loads Admin Tools feature */
         include_once $this->path( 'lib/class_admin_tools.php', 'dir' );
-        /** Loads export functionality */
-        include_once $this->path( 'lib/class_property_export.php', 'dir' );
         /** Loads all the metaboxes for the property page */
         include_once $this->path( 'lib/class_core.php', 'dir' );
         /** Load set of static methods for mail notifications */
@@ -77,6 +75,9 @@ namespace UsabilityDynamics\WPP {
 
         //** Initiate AJAX Handler */
         new Ajax();
+
+        //** Handles Export (XML/JSON/CSV) functionality */
+        new Export();
 
         //** Initiate Admin UI */
         if( is_admin() ) {